Module Uring.Linkat_flags

include FLAGS

type t = private int

A set of flags.

val empty : t

val of_int : int -> t

val (+) : t -> t -> t

a + b is the union of the sets.

val mem : t -> t -> bool

mem x flags is true iff x is a subset of flags.

val empty_path : t

If the old path is empty, link to old_dir_fd.

val symlink_follow : t

If the old path is a symlink, link its target.