Module Uring.Statx

type t

A statx struct.

type kind = [

| `Unknown

| `Fifo

| `Character_special

| `Directory

| `Block_device

| `Regular_file

| `Symbolic_link

| `Socket

]

val pp_kind : kind Fmt.t

val create : unit -> t

Use create to make a statx result buffer to pass to statx.

module Flags : sig ... end

module Attr : sig ... end

module Mask : sig ... end

You may wish to use Mask.check to verify the field has actually been returned with a sensible value first.

val blksize : t -> Stdlib.Int64.t

val attributes : t -> Stdlib.Int64.t

val nlink : t -> Stdlib.Int64.t

val uid : t -> Stdlib.Int64.t

val gid : t -> Stdlib.Int64.t

val ino : t -> Stdlib.Int64.t

val size : t -> Stdlib.Int64.t

val blocks : t -> Stdlib.Int64.t

val attributes_mask : t -> Stdlib.Int64.t

val rdev : t -> Stdlib.Int64.t

val dev : t -> Stdlib.Int64.t

val mask : t -> Stdlib.Int64.t

val mnt_id : t -> Stdlib.Int64.t

See Mask.mnt_id.

val dio_mem_align : t -> Stdlib.Int64.t

See Mask.dioalign.

val dio_offset_align : t -> Stdlib.Int64.t

See Mask.dioalign.

val atime_sec : t -> int64

val btime_sec : t -> int64

val ctime_sec : t -> int64

val mtime_sec : t -> int64

val atime_nsec : t -> int

val btime_nsec : t -> int

val ctime_nsec : t -> int

val mtime_nsec : t -> int

val mode : t -> int

val perm : t -> int

val kind : t -> kind