Module Statx.Flags

include FLAGS

type t = private int

A set of flags.

val empty : t

val of_int : int -> t

val (+) : t -> t -> t

a + b is the union of the sets.

val mem : t -> t -> bool

mem x flags is true iff x is a subset of flags.

val empty_path : t

empty_path signals that if the pathname is an empty string then operate on the file referred to by the fd (which can refer to any type of file). If fd is not specified then the call operates on the current working directory.

val no_automount : t

no_automount signals that statx should not automount the basename component of the path if it is an automount point. This can be used in tools that scan directories to prevent mass-automounting of a directory of automount points.

val symlink_nofollow : t

symlink_nofollow signals that if the path is a symbolic link, then return information about the link itself.

val statx_sync_as_stat : t

statx_sync_as_stat is the filesystem-specific behaviour in response to stat calls.

val statx_force_sync : t

statx_force_sync forces synchronisation with the server, if the filesystem is a network-backed one.

val statx_dont_sync : t

statx_dont_sync signals that locally cached timestamps are sufficient, if run on a network-backed filesystem.