Module Statx.Mask

include FLAGS

type t = private int

A set of flags.

val of_int : int -> t

val (+) : t -> t -> t

a + b is the union of the sets.

val mem : t -> t -> bool

mem x flags is true iff x is a subset of flags.

val type' : t

val mode : t

val nlink : t

val uid : t

val gid : t

val atime : t

val mtime : t

val ctime : t

val ino : t

val size : t

val blocks : t

val basic_stats : t

All of the above flags.

val btime : t

val mnt_id : t

As of Linux 5.8

val dioalign : t

As of Linux 6.1

val check : Stdlib.Int64.t -> t -> bool

check mask t checks if t is set in mask.