Module Uring.Msghdr

type t

val create : ?n_fds:int -> ?addr:Sockaddr.t -> Cstruct.t list -> t

create buffs makes a new msghdr using the buffs for the underlying iovec.

Requires List.length buffs <= Uring.iov_max

parameter addr
The remote address. Use Sockaddr.create to create a dummy address that will be filled when data is received.

parameter n_fds
Reserve space to receive this many FDs (default 0)

val get_fds : t -> Unix.file_descr list